Store fetched tiles in a cache; remove them when the bucket is assembled
This avoids repeated downloads of the same tiles when a server or
network error causes the script to abort before all tiles for the
specified bucket have been downloaded.

The tiles are stored in a provider-specific subdirectory of:

  - for non-Windows systems:

      '$XDG_CACHE_HOME/FlightGear photoscenery creator'
      ('$HOME/.cache/FlightGear photoscenery creator' if XDG_CACHE_HOME
      is empty or unset);

  - for Windows: '%APPDATA%/FlightGear photoscenery creator/cache'.

(this can be overridden with option --cache-dir). The tiles are
automatically deleted when the bucket image they were downloaded for is
assembled. Every file that remains in the cache is complete (each tile
is downloaded to a temporary file which is renamed when the download is
finished and removed if it's aborted, so that we don't keep any
incomplete file in cache).