Add option --clear-cache
Using --clear-cache=PROVIDER clears the tile cache for PROVIDER. If
PROVIDER is the special value 'ALL', then all provider-specific subdirs
are cleared. The option may be used with none of --lat, --lon and
--index if one only wants to clear the tile cache. It is also possible
to combine the clearing and downloading operations in a single command,
in which case cache clearing happens first.